requirements.txt

aiohttp
aiodns
aiosqlite
ccxt
cryptography
fastapi
httpx
mcp
numpy
pandas
pydantic
python-dateutil
python-dotenv
pytz
requests
uvicorn
websockets
pytest
pytest-asyncio